Restructure main logic
