+++ src/main/java/voyatrip/ui/Message.java
@@ -1,11 +1,9 @@
 package voyatrip.ui;
 
 public class Message {
+    // user message
+    static final String WELCOME_MESSAGE = "Welcome to VoyaTrip! How can I help you today?";
+    static final String GOODBYE_MESSAGE = "Goodbye! Wish you have a lovely trip!";
     // exception
-    private static final String INVALID_COMMAND_MESSAGE = "Invalid command. Please try again.";
-
-    // exception
-    public static void printInvalidCommand() {
-        System.out.println(INVALID_COMMAND_MESSAGE);
-    }
+    static final String INVALID_COMMAND_MESSAGE = "Invalid command. Please try again.";
 }
